requirements.txt
Pillow>=10.0.0
colormath>=3.0.0
nbt